perf(imaging): cut ~640 MB from the per-frame memory peak
The 12.6 MP ASI676MC frame (3552x3552) was being converted to float
several times per capture at full resolution, spiking RSS to ~1.9 GB
while imaging and leaving a ~1.2 GB working-set high-water mark at idle.

Four full-frame temporaries removed, all behaviour-preserving:

- image_processor: the display histogram built a float64 temp via
  (raw_array / 257) -- 303 MB. Bin the uint16 data directly over
  [0, 257*256) instead: bin width is exactly 257, so bin == floor(v/257),
  the identical bucketing, verified exhaustively over all 65536 values.
- ml_service: corner analysis took a float64 grayscale mean plus a second
  array for the normalisation. float32 accumulator, in-place divide, and
  overwrite_input on the final median (gray is dead after it).
- star_detection: FWHM grayscale float64 -> float32. Provably exact for
  integer input; the half-max arithmetic fits in a 24-bit mantissa.
- image_stretch: scale in place after astype instead of allocating the
  quotient, multiply in place before the uint8 cast, and drop the
  defensive copy in _normalize_channel_medians (its only caller owns a
  private frame and rebinds immediately).

Also stop caching RAW_RGB_NO_WB for reprocessing when RAW_RGB_16BIT is
present -- it is only ever read as the fallback for a missing 16-bit
array, so in RAW16 mode it pinned a full frame with no reader.

Measured with tracemalloc at the real 3552x3552 geometry: 298 + 133 + 48
+ 158 MB off the imaging peak, 38 MB off the idle floor. Verified
bit-exact against the previous implementations across 32 auto_stretch
branch combinations, and guarded against mutating the caller's raw_16bit
(which can be the camera's live ping-pong buffer).

+++ services/star_detection.py
@@ -90,10 +90,14 @@ def estimate_fwhm(image, stars, radius=8):
     if not stars or image is None or image.size == 0:
         return 0.0
 
+    # float32, not float64: FWHM is a half-max crossing measured on a small
+    # cutout, so ~7 significant digits is already far more precision than the
+    # measurement carries, and it halves a full-frame temp (50 MB vs 101 MB on
+    # a 3552x3552 frame).
     if len(image.shape) == 3:
-        gray = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY).astype(np.float64)
+        gray = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY).astype(np.float32)
     else:
-        gray = image.astype(np.float64)
+        gray = image.astype(np.float32)
 
     h, w = gray.shape
     fwhm_values = []
